+++ src/domain/bc/constant.rs
@@ -2,12 +2,12 @@ use crate::domain::bc::BCCheck;
 use crate::util::*;
 
 pub struct ConstantCheck<const GRID_DIMENSION: usize> {
-    value: f32,
+    value: f64,
     bound: AABB<GRID_DIMENSION>,
 }
 
 impl<const GRID_DIMENSION: usize> ConstantCheck<GRID_DIMENSION> {
-    pub fn new(value: f32, bound: AABB<GRID_DIMENSION>) -> Self {
+    pub fn new(value: f64, bound: AABB<GRID_DIMENSION>) -> Self {
         ConstantCheck { value, bound }
     }
 }
@@ -15,7 +15,7 @@ impl<const GRID_DIMENSION: usize> ConstantCheck<GRID_DIMENSION> {
 impl<const GRID_DIMENSION: usize> BCCheck<GRID_DIMENSION>
     for ConstantCheck<GRID_DIMENSION>
 {
-    fn check(&self, coord: &Coord<GRID_DIMENSION>) -> Option<f32> {
+    fn check(&self, coord: &Coord<GRID_DIMENSION>) -> Option<f64> {
         if self.bound.contains(coord) {
             return None;
         }
@@ -35,7 +35,7 @@ mod unit_tests {
         let n_r = bound.buffer_size();
         let mut buffer = fftw::array::AlignedVec::new(n_r);
         for i in 0..n_r {
-            buffer.as_slice_mut()[i] = i as f32;
+            buffer.as_slice_mut()[i] = i as f64;
         }
         let bc = ConstantCheck::new(-1.0, bound);
         for i in 0..n_r {
@@ -46,13 +46,13 @@ mod unit_tests {
         {
             let v = bc.check(&vector![-1]);
             assert!(v.is_some());
-            assert_approx_eq!(f32, v.unwrap(), -1.0);
+            assert_approx_eq!(f64, v.unwrap(), -1.0);
         }
 
         {
             let v = bc.check(&vector![11]);
             assert!(v.is_some());
-            assert_approx_eq!(f32, v.unwrap(), -1.0);
+            assert_approx_eq!(f64, v.unwrap(), -1.0);
         }
     }
 }
